+++ plugins/inventory/esxi_hosts.py
@@ -39,14 +39,16 @@ EXAMPLES = r"""
 
 
 # Simple configuration with in-file authentication parameters
+---
 plugin: vmware.vmware.esxi_hosts
 hostname: 10.65.223.31
 username: administrator@vsphere.local
 password: Esxi@123$%
 validate_certs: false
-
+...
 
 # More complex configuration. Authentication parameters are assumed to be set as environment variables.
+---
 plugin: vmware.vmware.esxi_hosts
 
 # Create groups based on host paths
@@ -57,6 +59,10 @@ properties: ["name", "capability"]
 groups:
   vmotion_supported: capability.vmotionSupported
 
+# Filter out hosts using jinja patterns. For example, filter out powered off hosts
+filter_expressions:
+  - 'summary.runtime.powerState == "poweredOff"'
+
 # Only gather hosts found in certain paths
 search_paths:
   - /DC1/host/ClusterA
@@ -64,6 +70,8 @@ search_paths:
   - /DC3
 
 # Set custom inventory hostnames based on attributes
+# If more than one host has the same name, only the first host is shown in the inventory and a warning is thrown.
+# If strict is true, this warning is considered a fatal error.
 hostnames:
   - "'ESXi - ' + name + ' - ' + management_ip"
   - "'ESXi - ' + name"
@@ -75,6 +83,7 @@ compose:
   # assuming path is something like /MyDC/host/MyCluster
   datacenter: "(path | split('/'))[1]"
   cluster: "(path | split('/'))[3]"
+...
 """
 
 try:
@@ -193,8 +202,13 @@ class InventoryModule(VmwareInventoryBase):
         if "name" not in properties_param:
             properties_param.append("name")
 
+        # needed to filter out disconnected or unreachable hosts in self.populate_from_vcenter
         if "summary.runtime.connectionState" not in properties_param:
             properties_param.append("summary.runtime.connectionState")
+
+        # needed by keyed_groups default
+        if "summary.runtime.powerState" not in properties_param:
+            properties_param.append("summary.runtime.powerState")
 
         return properties_param
 
@@ -233,29 +247,18 @@ class InventoryModule(VmwareInventoryBase):
                 self.add_tags_to_object_properties(esxi_host)
 
             self.set_inventory_hostname(esxi_host)
-            if esxi_host.inventory_hostname not in hostvars:
-                hostvars[esxi_host.inventory_hostname] = esxi_host.properties
-                self.__update_inventory(esxi_host)
+            self.add_host_object_from_vcenter_to_inventory(new_host=esxi_host, hostvars=hostvars)
 
         return hostvars
 
-    def __update_inventory(self, esxi_host):
-        self.add_host_to_inventory(esxi_host)
-        self.add_host_to_groups_based_on_path(esxi_host)
-        self.set_host_variables_from_host_properties(esxi_host)
-
-    def add_host_to_inventory(self, esxi_host: EsxiInventoryHost):
+    def set_default_ansible_host_var(self, vmware_host_object):
         """
-        Add the host to the inventory and any groups that the user wants to create based on inventory
-        parameters like groups or keyed groups.
+            Sets the default ansible_host var. This is usually an IP that is dependent on the object type.
+            This is a default because the user can override this via compose
+            Args:
+              vmware_host_object: EsxiInventoryHost, The host object that should be used
         """
-        strict = self.get_option("strict")
-        self.inventory.add_host(esxi_host.inventory_hostname)
-        self.inventory.set_variable(esxi_host.inventory_hostname, "ansible_host", esxi_host.management_ip)
-
-        self._set_composite_vars(
-            self.get_option("compose"), esxi_host.properties, esxi_host.inventory_hostname, strict=strict)
-        self._add_host_to_composed_groups(
-            self.get_option("groups"), esxi_host.properties, esxi_host.inventory_hostname, strict=strict)
-        self._add_host_to_keyed_groups(
-            self.get_option("keyed_groups"), esxi_host.properties, esxi_host.inventory_hostname, strict=strict)
+        self.inventory.set_variable(
+            vmware_host_object.inventory_hostname, "ansible_host",
+            vmware_host_object.management_ip
+        )