+++ plugins/inventory/vms.py
@@ -48,14 +48,16 @@ EXAMPLES = r"""
 
 
 # Simple configuration with in-file authentication parameters
+---
 plugin: vmware.vmware.vms
 hostname: 10.65.223.31
 username: administrator@vsphere.local
 password: Esxi@123$%
 validate_certs: false
-
+...
 
 # More complex configuration. Authentication parameters are assumed to be set as environment variables.
+---
 plugin: vmware.vmware.vms
 # Create groups based on host paths
 group_by_paths: true
@@ -74,7 +76,12 @@ search_paths:
   - /DC1/vm/ClusterA
   - /DC1/vm/ClusterC
   - /DC3
+# Filter out VMs using jinja patterns. For example, filter out powered off VMs
+filter_expressions:
+  - 'summary.runtime.powerState == "poweredOff"'
 # Set custom inventory hostnames based on attributes
+# If more than one host has the same name, only the first host is shown in the inventory and a warning is thrown.
+# If strict is true, this warning is considered a fatal error.
 hostnames:
   - "'VM - ' + name + ' - ' + guest.ipAddress"
   - "'VM - ' + name + ' - ' + config.instanceUuid"
@@ -85,17 +92,20 @@ compose:
   # assuming path is something like /MyDC/vms/MyCluster
   datacenter: "(path | split('/'))[1]"
   cluster: "(path | split('/'))[3]"
-
+...
 
 # Use Tags and Tag Categories to create groups
+---
 plugin: vmware.vmware.vms
 gather_tags: true
 keyed_groups:
   - key: tags_by_category.OS
     prefix: "vmware_tag_os_category_"
     separator: ""
+...
 
 # customizing hostnames based on VM's FQDN. The second hostnames template acts as a fallback mechanism.
+---
 plugin: vmware.vmware.vms
 hostnames:
   - 'config.name+"."+guest.ipStack.0.dnsConfig.domainName'
@@ -107,8 +117,10 @@ properties:
   - 'guest.ipAddress'
   - 'guest.guestFamily'
   - 'guest.ipStack'
+...
 
 # Select a specific IP address for use by ansible when multiple NICs are present on the VM
+---
 plugin: vmware.vmware.vms
 compose:
   # Set the IP address used by ansible to one that starts by 10.42. or 10.43.
@@ -120,16 +132,19 @@ compose:
     | select('match', '^10.42.*|^10.43.*')
     | list
     | first
-  properties:
-    - guest.net
+properties:
+  - guest.net
+...
 
 # Group hosts using Jinja2 conditionals
+---
 plugin: vmware.vmware.vms
 properties:
   - 'config.datastoreUrl'
 groups:
   slow_storage: "'Nas01' in config.datastoreUrl[0].name"
   fast_storage: "'SSD' in config.datastoreUrl[0].name"
+...
 """
 
 try:
@@ -222,6 +237,11 @@ class InventoryModule(VmwareInventoryBase):
         if "name" not in properties_param:
             properties_param.append("name")
 
+        # needed by keyed_groups default value
+        if "config.guestId" not in properties_param:
+            properties_param.append("config.guestId")
+
+        # needed by keyed_groups default value
         if "summary.runtime.powerState" not in properties_param:
             properties_param.append("summary.runtime.powerState")
 
@@ -259,29 +279,18 @@ class InventoryModule(VmwareInventoryBase):
                 self.add_tags_to_object_properties(vm)
 
             self.set_inventory_hostname(vm)
-            if vm.inventory_hostname not in hostvars:
-                hostvars[vm.inventory_hostname] = vm.properties
-                self.__update_inventory(vm)
+            self.add_host_object_from_vcenter_to_inventory(new_host=vm, hostvars=hostvars)
 
         return hostvars
 
-    def __update_inventory(self, vm):
-        self.add_host_to_inventory(vm)
-        self.add_host_to_groups_based_on_path(vm)
-        self.set_host_variables_from_host_properties(vm)
-
-    def add_host_to_inventory(self, vm: VmInventoryHost):
+    def set_default_ansible_host_var(self, vmware_host_object):
         """
-        Add the host to the inventory and any groups that the user wants to create based on inventory
-        parameters like groups or keyed groups.
+            Sets the default ansible_host var. This is usually an IP that is dependent on the object type.
+            This is a default because the user can override this via compose
+            Args:
+              vmware_host_object: EsxiInventoryHost, The host object that should be used
         """
-        strict = self.get_option("strict")
-        self.inventory.add_host(vm.inventory_hostname)
-        self.inventory.set_variable(vm.inventory_hostname, "ansible_host", vm.guest_ip)
-
-        self._set_composite_vars(
-            self.get_option("compose"), vm.properties, vm.inventory_hostname, strict=strict)
-        self._add_host_to_composed_groups(
-            self.get_option("groups"), vm.properties, vm.inventory_hostname, strict=strict)
-        self._add_host_to_keyed_groups(
-            self.get_option("keyed_groups"), vm.properties, vm.inventory_hostname, strict=strict)
+        self.inventory.set_variable(
+            vmware_host_object.inventory_hostname, "ansible_host",
+            vmware_host_object.guest_ip
+        )