+++ plugins/inventory_utils/_base.py
@@ -13,17 +13,20 @@ from ansible.plugins.inventory import BaseInventoryPlugin, Constructable, Cachea
 from ansible.parsing.yaml.objects import AnsibleVaultEncryptedUnicode
 from ansible.module_utils.common.text.converters import to_native
 from ansible.module_utils.common.dict_transformations import camel_dict_to_snake_dict
+from ansible.utils.display import Display
 
-
-from ansible_collections.vmware.vmware.plugins.module_utils.clients._pyvmomi import PyvmomiClient
-from ansible_collections.vmware.vmware.plugins.module_utils.clients._rest import VmwareRestClient
-from ansible_collections.vmware.vmware.plugins.module_utils._vmware_folder_paths import (
+from ansible_collections.vmware.vmware.plugins.module_utils.clients.pyvmomi import PyvmomiClient
+from ansible_collections.vmware.vmware.plugins.module_utils.clients.rest import VmwareRestClient
+from ansible_collections.vmware.vmware.plugins.module_utils._folder_paths import (
     get_folder_path_of_vsphere_object
 )
-from ansible_collections.vmware.vmware.plugins.module_utils._vmware_facts import (
+from ansible_collections.vmware.vmware.plugins.module_utils._facts import (
     vmware_obj_to_json,
     flatten_dict
 )
+
+
+DISPLAY = Display()
 
 
 class VmwareInventoryHost(ABC):
@@ -67,6 +70,7 @@ class VmwareInventoryHost(ABC):
     def get_properties_from_pyvmomi(self, properties_to_gather, pyvmomi_client):
         properties = vmware_obj_to_json(self.object, properties_to_gather)
         properties['path'] = self.path
+        properties['moid'] = self.object._GetMoId()
 
         # Custom values
         if hasattr(self.object, "customValue"):
@@ -98,15 +102,15 @@ class VmwareInventoryBase(BaseInventoryPlugin, Constructable, Cacheable):
         username, password = self.get_credentials_from_options()
 
         try:
-            self.pyvmomi_client = PyvmomiClient({
-                'hostname': self.get_option("hostname"),
-                'username': username,
-                'password': password,
-                'port': self.get_option("port"),
-                'validate_certs': self.get_option("validate_certs"),
-                'http_proxy_host': self.get_option("proxy_host"),
-                'http_proxy_port': self.get_option("proxy_port")
-            })
+            self.pyvmomi_client = PyvmomiClient(
+                hostname=self.get_option("hostname"),
+                username=username,
+                password=password,
+                port=self.get_option("port"),
+                validate_certs=self.get_option("validate_certs"),
+                http_proxy_host=self.get_option("proxy_host"),
+                http_proxy_port=self.get_option("proxy_port")
+            )
         except Exception as e:
             raise AnsibleParserError(message=to_native(e))
 
@@ -120,16 +124,16 @@ class VmwareInventoryBase(BaseInventoryPlugin, Constructable, Cacheable):
         username, password = self.get_credentials_from_options()
 
         try:
-            self.rest_client = VmwareRestClient({
-                'hostname': self.get_option("hostname"),
-                'username': username,
-                'password': password,
-                'port': self.get_option("port"),
-                'validate_certs': self.get_option("validate_certs"),
-                'http_proxy_host': self.get_option("proxy_host"),
-                'http_proxy_port': self.get_option("proxy_port"),
-                'http_proxy_protocol': self.get_option("proxy_protocol")
-            })
+            self.rest_client = VmwareRestClient(
+                hostname=self.get_option("hostname"),
+                username=username,
+                password=password,
+                port=self.get_option("port"),
+                validate_certs=self.get_option("validate_certs"),
+                http_proxy_host=self.get_option("proxy_host"),
+                http_proxy_port=self.get_option("proxy_port"),
+                http_proxy_protocol=self.get_option("proxy_protocol")
+            )
         except Exception as e:
             raise AnsibleParserError(message=to_native(e))
 
@@ -325,3 +329,81 @@ class VmwareInventoryBase(BaseInventoryPlugin, Constructable, Cacheable):
 
         if last_created_group:
             self.inventory.add_host(vmware_host_object.inventory_hostname, last_created_group)
+
+    def host_should_be_filtered_out(self, vmware_host_object):
+        """
+            Returns true if the provided host and properties cause any of the filter expressions
+            to evaluate to true. This indicates that the host should be removed from the final
+            inventory.
+            Returns false otherwise.
+        """
+        for jinja_expression in self.get_option('filter_expressions'):
+            try:
+                if self._compose(jinja_expression, vmware_host_object.properties):
+                    return True
+            except Exception as e:  # pylint: disable=broad-except
+                if self.get_option("strict"):
+                    raise AnsibleError(
+                        "Could not evaluate %s as host filter - %s" % (jinja_expression, to_native(e))
+                    )
+
+        return False
+
+    def _handle_duplicate_host(self, existing_host_vars, new_host):
+        """
+            Handles the instance where two hosts have the same inventory hostname. If the user has strict set to true, this
+            should be an error. Otherwise it is just a warning.
+            Args:
+                existing_host_vars: dict, The host properties of the existing host (for example, hostvars[inventory_hostname])
+                new_host: VmwareInventoryHost, The host object that was going to be added, but is a duplicate hostname of another
+        """
+        if self.get_option('strict'):
+            raise AnsibleError(
+                "Host %s has the same inventory hostname (%s) as %s. This is a fatal issue since strict is true." %
+                (existing_host_vars['moid'], new_host.inventory_hostname, new_host.object._GetMoId())
+            )
+        else:
+            DISPLAY.warning(
+                "Host %s has the same inventory hostname (%s) as %s. Only the first host, %s, will be included since strict is false." %
+                (existing_host_vars['moid'], new_host.inventory_hostname, new_host.object._GetMoId(), existing_host_vars['moid'])
+            )
+
+    def add_host_object_from_vcenter_to_inventory(self, new_host, hostvars):
+        """
+            Add a new host to the inventory and populate host vars based on object properties.
+            If the host should be filtered out or if a host with the same name already exists, this method will handle that
+            and simply return without processing the host.
+            Args:
+                new_host: VmwareInventoryHost, The new host to add to the inventory
+                hostvars: dict, The hostvars dict that should be updated with the new host's properties
+        """
+        if new_host.inventory_hostname in hostvars:
+            self._handle_duplicate_host(hostvars[new_host.inventory_hostname], new_host)
+            return
+
+        if self.host_should_be_filtered_out(new_host):
+            return
+
+        hostvars[new_host.inventory_hostname] = new_host.properties
+        self.inventory.add_host(new_host.inventory_hostname)
+        self.set_default_ansible_host_var(new_host)
+
+        _strict = self.get_option("strict")
+        self._set_composite_vars(
+            self.get_option("compose"), new_host.properties, new_host.inventory_hostname, strict=_strict)
+        self._add_host_to_composed_groups(
+            self.get_option("groups"), new_host.properties, new_host.inventory_hostname, strict=_strict)
+        self._add_host_to_keyed_groups(
+            self.get_option("keyed_groups"), new_host.properties, new_host.inventory_hostname, strict=_strict)
+
+        self.add_host_to_groups_based_on_path(new_host)
+        self.set_host_variables_from_host_properties(new_host)
+
+    def set_default_ansible_host_var(self, vmware_host_object):
+        """
+            Sets the default ansible_host var. This is usually an IP that is dependent on the object type.
+            This is a default because the user can override this via compose
+            Args:
+              vmware_host_object: The host object that should be used. The type will be dependent on the plugin type.
+        """
+        raise NotImplementedError('ansible_host should be defined in the inventory plugin class.')