+++ plugins/module_utils/_module_pyvmomi_base.py
@@ -15,12 +15,13 @@ except ImportError:
     pass
     # handled in base class
 
-from ansible_collections.vmware.vmware.plugins.module_utils.clients._pyvmomi import PyvmomiClient
+from ansible_collections.vmware.vmware.plugins.module_utils.clients.pyvmomi import PyvmomiClient
+from ansible_collections.vmware.vmware.plugins.module_utils._folder_paths import format_folder_path_as_vm_fq_path
 
 
 class ModulePyvmomiBase(PyvmomiClient):
     def __init__(self, module):
-        super().__init__(module.params)
+        super().__init__(**module.params)
         self.module = module
         self.params = module.params
 
@@ -58,25 +59,30 @@ class ModulePyvmomiBase(PyvmomiClient):
         Returns:
             list(object) or list() if no matches are found
         """
+        identifier = name
         if not search_root_folder:
             search_root_folder = self.content.rootFolder
+        if isinstance(vimtype, list):
+            vimtype = vimtype[0]
 
-        obj = list()
-        container = self.content.viewManager.CreateContainerView(
-            search_root_folder, vimtype, True)
+        results = []
+        for managed_object_ref in self.get_managed_object_references(vimtype, properties=['name'], folder=search_root_folder):
+            if len(results) > 0 and not return_all:
+                break
 
-        for c in container.view:
-            if name in [c.name, c._GetMoId()]:
-                if return_all is False:
-                    return c
-                else:
-                    obj.append(c)
+            obj_skeleton = managed_object_ref.obj
+            if obj_skeleton._GetMoId() == identifier:
+                vim_obj = self.create_vim_object_from_moid(obj_skeleton._GetMoId(), vimtype)
+                results.append(vim_obj)
+                continue
 
-        if len(obj) > 0:
-            return obj
-        else:
-            # for backwards-compat
-            return None
+            for property in managed_object_ref.propSet:
+                if property.name == "name" and property.val == identifier:
+                    vim_obj = self.create_vim_object_from_moid(obj_skeleton._GetMoId(), vimtype)
+                    results.append(vim_obj)
+                    break
+
+        return results
 
     def get_standard_portgroup_by_name_or_moid(self, identifier, fail_on_missing=False):
         """
@@ -88,9 +94,11 @@ class ModulePyvmomiBase(PyvmomiClient):
             The standard portgroup object
         """
         pg = self.get_objs_by_name_or_moid([vim.Network], identifier)
-        if not pg and fail_on_missing:
+        if pg:
+            return pg[0]
+        if fail_on_missing:
             self.module.fail_json(msg="Unable to find standard portgroup with name or MOID %s" % identifier)
-        return pg
+        return None
 
     def get_dvs_portgroup_by_name_or_moid(self, identifier, fail_on_missing=False):
         """
@@ -102,13 +110,15 @@ class ModulePyvmomiBase(PyvmomiClient):
             The distributed portgroup object
         """
         pg = self.get_objs_by_name_or_moid([vim.dvs.DistributedVirtualPortgroup], identifier)
-        if not pg and fail_on_missing:
+        if pg:
+            return pg[0]
+        if fail_on_missing:
             self.module.fail_json(msg="Unable to find distributed portgroup with name or MOID %s" % identifier)
-        return pg
+        return None
 
-    def get_vm_using_params(
+    def get_vms_using_params(
             self, name_param='name', uuid_param='uuid', moid_param='moid', fail_on_missing=False,
-            name_match_param='name_match', use_instance_uuid_param='use_instance_uuid'):
+            name_match_param='name_match', use_instance_uuid_param='use_instance_uuid', folder_param='folder'):
         """
             Get the vms matching the common module params related to vm identification: name, uuid, or moid. Since
             MOID and UUID are unique identifiers, they are tried first. If they are not set, a search by name is tried
@@ -122,30 +132,27 @@ class ModulePyvmomiBase(PyvmomiClient):
                 name_match_param: Set the parameter key that corredsponds to the name_match option
                 use_instance_uuid_param: Set the parameter key that corredsponds use_instance_uuid option
                 fail_on_missing: If true, an error will be thrown if no VMs are found
+                folder_param: Set the parameter key that corresponds to the folder that contains the VM
             Returns:
-                list(vm), or None if no matches were found
+                list(vm). In most cases a list of length 1 but when searching by name, you can get multiple matches.
         """
-        if self.params.get(moid_param):
-            _search_type, _search_id, _search_value = 'moid', moid_param, self.params.get(moid_param)
-        elif self.params.get(uuid_param):
-            _search_type, _search_id, _search_value = 'uuid', uuid_param, self.params.get(uuid_param)
-        elif self.params.get(name_param):
-            _search_type, _search_id, _search_value = 'name', name_param, self.params.get(name_param)
-        else:
-            if fail_on_missing:
-                self.module.fail_json(msg="Could not find any supported VM identifier params (name, uuid, or moid)")
-            else:
-                return None
+        _search_type, _search_id = self.__determine_search_param_type_and_id(
+            name_param=name_param, uuid_param=uuid_param, moid_param=moid_param, fail_on_missing=fail_on_missing
+        )
+        if not _search_type:
+            return []
 
         if _search_type == 'uuid':
-            _vm = self.si.content.searchIndex.FindByUuid(
-                instanceUuid=self.params.get(use_instance_uuid_param, True),
-                uuid=_search_value,
-                vmSearch=True
-            )
-            vms = [_vm] if _vm else None
+            vms = [self.si.content.searchIndex.FindByUuid(
+                instanceUuid=self.params.get(use_instance_uuid_param, True), uuid=self.params.get(_search_id), vmSearch=True
+            )]
+
         else:
-            vms = self.get_objs_by_name_or_moid([vim.VirtualMachine], _search_value, return_all=True)
+            folder = None
+            if self.params.get(folder_param):
+                _fq_path = format_folder_path_as_vm_fq_path(self.params.get(folder_param), self.params.get('datacenter'))
+                folder = self.get_folder_by_absolute_path(_fq_path, fail_on_missing=fail_on_missing)
+            vms = self.get_objs_by_name_or_moid([vim.VirtualMachine], self.params.get(_search_id), return_all=True, search_root_folder=folder)
 
         if vms and _search_type == 'name' and self.params.get(name_match_param):
             if self.params.get(name_match_param) == 'first':
@@ -156,9 +163,31 @@ class ModulePyvmomiBase(PyvmomiClient):
                 self.module.fail_json(msg="Unrecognized name_match option '%s' " % self.params.get(name_match_param))
 
         if not vms and fail_on_missing:
-            self.module.fail_json(msg="Unable to find VM with %s %s" % (_search_id, _search_value))
+            self.module.fail_json(msg="Unable to find VM with %s %s" % (_search_id, self.params.get(_search_id)))
 
         return vms
+
+    def __determine_search_param_type_and_id(self, name_param, uuid_param, moid_param, fail_on_missing):
+        """
+        Helper function for get_vms_using_params. Tries to determine which of the VM identifying parameters should be
+        used to search for matching VMs. Optionally throws an error if none of the params are valid.
+        Returns:
+          str, str: The type of identifier that was selected, and the parameter name for that identifier
+        """
+        if self.params.get(moid_param):
+            return 'moid', moid_param
+        elif self.params.get(uuid_param):
+            return 'uuid', uuid_param
+        elif self.params.get(name_param):
+            return 'name', name_param
+
+        if fail_on_missing:
+            self.module.fail_json(msg=(
+                "Could not find any supported VM identifier params (%s, %s, or %s)" %
+                (name_param, uuid_param, moid_param)
+            ))
+
+        return None, None
 
     def get_folders_by_name_or_moid(self, identifier, fail_on_missing=False):
         """
@@ -203,9 +232,11 @@ class ModulePyvmomiBase(PyvmomiClient):
                 datastore object or None
         """
         ds = self.get_objs_by_name_or_moid([vim.Datastore], identifier)
-        if not ds and fail_on_missing:
+        if ds:
+            return ds[0]
+        if fail_on_missing:
             self.module.fail_json(msg="Unable to find datastore with name or MOID %s" % identifier)
-        return ds
+        return None
 
     def get_datastore_cluster_by_name_or_moid(self, identifier, fail_on_missing=False, datacenter=None):
         """
@@ -222,7 +253,7 @@ class ModulePyvmomiBase(PyvmomiClient):
         """
         search_folder = None
         if datacenter and hasattr(datacenter, 'datastoreFolder'):
-            search_folder = datacenter.hostFolder
+            search_folder = datacenter.datastoreFolder
 
         data_store_cluster = self.get_objs_by_name_or_moid(
             [vim.StoragePod],
@@ -231,10 +262,12 @@ class ModulePyvmomiBase(PyvmomiClient):
             search_root_folder=search_folder
         )
 
-        if not data_store_cluster and fail_on_missing:
+        if data_store_cluster:
+            return data_store_cluster[0]
+        if fail_on_missing:
             self.module.fail_json(msg="Unable to find datastore cluster with name or MOID %s" % identifier)
 
-        return data_store_cluster
+        return None
 
     def get_resource_pool_by_name_or_moid(self, identifier, fail_on_missing=False):
         """
@@ -247,9 +280,11 @@ class ModulePyvmomiBase(PyvmomiClient):
                 resource pool object or None
         """
         pool = self.get_objs_by_name_or_moid([vim.ResourcePool], identifier)
-        if not pool and fail_on_missing:
+        if pool:
+            return pool[0]
+        if fail_on_missing:
             self.module.fail_json(msg="Unable to find resource pool with name %s" % identifier)
-        return pool
+        return None
 
     def get_all_vms(self, folder=None, recurse=True):
         """
@@ -275,9 +310,11 @@ class ModulePyvmomiBase(PyvmomiClient):
                 datacenter object or None
         """
         ds = self.get_objs_by_name_or_moid([vim.Datacenter], identifier)
-        if not ds and fail_on_missing:
+        if ds:
+            return ds[0]
+        if fail_on_missing:
             self.module.fail_json(msg="Unable to find datacenter with name or MOID %s" % identifier)
-        return ds
+        return None
 
     def get_cluster_by_name_or_moid(self, identifier, fail_on_missing=False, datacenter=None):
         """
@@ -301,11 +338,13 @@ class ModulePyvmomiBase(PyvmomiClient):
             return_all=False,
             search_root_folder=search_folder
         )
+        if cluster:
+            return cluster[0]
 
-        if not cluster and fail_on_missing:
+        if fail_on_missing:
             self.module.fail_json(msg="Unable to find cluster with name or MOID %s" % identifier)
 
-        return cluster
+        return None
 
     def get_esxi_host_by_name_or_moid(self, identifier, fail_on_missing=False):
         """
@@ -322,11 +361,12 @@ class ModulePyvmomiBase(PyvmomiClient):
             identifier,
             return_all=False,
         )
-
-        if not esxi_host and fail_on_missing:
+        if esxi_host:
+            return esxi_host[0]
+        if fail_on_missing:
             self.module.fail_json(msg="Unable to find ESXi host with name or MOID %s" % identifier)
 
-        return esxi_host
+        return None
 
     def get_sdrs_recommended_datastore_from_ds_cluster(self, ds_cluster):
         """
@@ -365,9 +405,12 @@ class ModulePyvmomiBase(PyvmomiClient):
         datastore = None
         datastore_freespace = 0
         for ds in datastores:
-            if isinstance(ds, vim.Datastore) and ds.summary.freeSpace > datastore_freespace:
-                if ds.summary.maintenanceMode == 'normal' and ds.summary.accessible:
-                    datastore = ds
-                    datastore_freespace = ds.summary.freeSpace
+            try:
+                if ds.summary.freeSpace > datastore_freespace:
+                    if ds.summary.maintenanceMode == 'normal' and ds.summary.accessible:
+                        datastore = ds
+                        datastore_freespace = ds.summary.freeSpace
+            except AttributeError:
+                continue
 
         return datastore