+++ plugins/module_utils/_module_rest_base.py
@@ -21,12 +21,12 @@ except ImportError:
     pass
     # handled in base class
 
-from ansible_collections.vmware.vmware.plugins.module_utils.clients._rest import VmwareRestClient
+from ansible_collections.vmware.vmware.plugins.module_utils.clients.rest import VmwareRestClient
 
 
 class ModuleRestBase(VmwareRestClient):
     def __init__(self, module):
-        super().__init__(connection_params=module.params)
+        super().__init__(**module.params)
         self.module = module
         self.params = module.params